backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
mysql-connector-python==8.4.0
httpx==0.27.0
pydantic==2.7.0
python-dotenv==1.0.1

bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

sqlalchemy==2.0.28
pymysql==1.1.0
cryptography==42.0.5
slowapi==0.1.9
